+++ server/src/main/scala/chess/domain/Square.scala
@@ -1,4 +1,0 @@
-package com.chessonline
-package chess.domain
-
-final case class Square(pieceOption: Option[Piece])