+++ server/src/test/scala/chess/domain/TestData.scala
@@ -22,8 +22,6 @@ object TestData {
 
   val whitePawn: Piece = Piece(White, Pawn)
   val blackPawn: Piece = Piece(Black, Pawn)
-  val whitePawnSquare: Square = Square(Some(whitePawn))
-  val blackPawnSquare: Square = Square(Some(blackPawn))
 
   val whiteKing: Piece = Piece(White, King)
   val blackKing: Piece = Piece(Black, King)
@@ -32,13 +30,9 @@ object TestData {
 
   val whiteRook: Piece = Piece(White, Rook)
   val blackRook: Piece = Piece(Black, Rook)
-  val whiteRookSquare: Square = Square(Some(whiteRook))
-  val blackRookSquare: Square = Square(Some(blackRook))
 
   val whiteBishop: Piece = Piece(White, Bishop)
   val blackBishop: Piece = Piece(White, Bishop)
-  val whiteBishopSquare: Square = Square(Some(whiteBishop))
-  val blackBishopSquare: Square = Square(Some(blackBishop))
 
   val whiteKnight: Piece = Piece(White, Knight)